core/chains: generic chains and nodes ORM (#6424)
* core/chains: generic chains and nodes ORM

* core/chains/evm: use generic ORM

* core/chains: dedupe and document ORMs

* docs

* more docs; prefix ChainType consts

+++ core/web/resolver/mutation.go
@@ -391,7 +391,7 @@ func (r *Resolver) DeleteNode(ctx context.Context, args struct {
 		return nil, err
 	}
 
-	err = r.App.EVMORM().DeleteNode(int64(id))
+	err = r.App.EVMORM().DeleteNode(id)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			// Sending the SQL error as the expected error to happen
